+++ Changelog.md
@@ -1,7 +1,7 @@
-- [ ] Make fixed scalar types default to int32 and int64
 - [ ] Honour [packed=false] flag.
 
 ## 0.9.1: Unreleased
+- [x] Make fixed scalar types default to int32 and int64
 - [x] Always serialize proto2 type fields
 - [x] Support default values (proto2)
 - [x] Support required (proto2) fields